# api_interface_demo/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Sitewise Interface API
  description: Sitewise Interface API Draft001
  version: 1.0.2

  # TODO:
  #  - finish responses and validation
  #  - testing
servers:
# - url: http://Api114partial-env.eba-wam2eszv.eu-west-1.elasticbeanstalk.com
- url: /
paths:
...
  /sites:
    get:
      tags:
      - Sites
      operationId: get_sites
      responses:
        "200":
          description: "{'ResponseMetadata': {'RequestId': '964b0463-0041-49dc-b62f-4f73a9d690cd',\
            \ \n'HTTPStatusCode': 200, \n'HTTPHeaders': {'date': 'Mon, 28 Dec 2020\
            \ 11:04:37 GMT', 'content-type': 'application/json', \n'content-length':\
            \ '406', 'connection': 'keep-alive', 'x-amzn-requestid': '964b0463-0041-49dc-b62f-4f73a9d690cd'},\
            \ \n'RetryAttempts': 0}, \n\n'assetSummaries': [\n\n{'id': '6627f48c-8d0a-481f-a305-b28e5258e319',\
            \ \n'arn': 'arn:aws:iotsitewise:eu-west-1:241091466505:asset/6627f48c-8d0a-481f-a305-b28e5258e319',\
            \ \n'name': 'site11', 'assetModelId': 'a106a0e4-1d20-46dc-a7a3-8dd3f902f48d',\
            \ \n'creationDate': datetime.datetime(2020, 12, 28, 10, 40, 35, tzinfo=tzlocal()),\
            \ \n'lastUpdateDate': datetime.datetime(2020, 12, 28, 10, 40, 35, tzinfo=tzlocal()),\
            \ \n'status': {'state': 'ACTIVE'}, 'hierarchies': [{'id': 'd883ac38-6437-4180-8ef5-124d0358241a',\
            \ \n'name': 'ZoneChilds'}]},\n\n{'id': '6627f48c-8d0a-481f-a305-b28e5258e319',\
            \ \n'arn': 'arn:aws:iotsitewise:eu-west-1:241091466505:asset/6627f48c-8d0a-481f-a305-b28e5258e319',\
            \ \n'name': 'site22', 'assetModelId': 'a106a0e4-1d20-46dc-a7a3-8dd3f902f48d',\
            \ \n'creationDate': datetime.datetime(2020, 12, 28, 10, 40, 35, tzinfo=tzlocal()),\
            \ \n'lastUpdateDate': datetime.datetime(2020, 12, 28, 10, 40, 35, tzinfo=tzlocal()),\
            \ \n'status': {'state': 'ACTIVE'}, 'hierarchies': [{'id': 'd883ac38-6437-4180-8ef5-124d0358241a',\
            \ \n'name': 'ZoneChilds'}]}\n]}\n"
      x-openapi-router-controller: swagger_server.controllers.sites_controller
    post:
      tags:
      - Sites
      operationId: create_site
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameObj'
        required: true
      responses:
        "201":
          description: "{'ResponseMetadata': {'RequestId': 'a7f4441a-204a-490f-8abc-d8225d1a9abc',\
            \ \n'HTTPStatusCode': 202, \n'HTTPHeaders': {'date': 'Mon, 28 Dec 2020\
            \ 09:40:35 GMT', 'content-type': 'application/json', \n'content-length':\
            \ '184', 'connection': 'keep-alive', 'x-amzn-requestid': 'a7f4441a-204a-490f-8abc-d8225d1a9abc'},\
            \ \n'RetryAttempts': 0}, \n'assetId': '6627f48c-8d0a-481f-a305-b28e5258e319',\
            \ \n'assetArn': 'arn:aws:iotsitewise:eu-west-1:241091466505:asset/6627f48c-8d0a-481f-a305-b28e5258e319',\
            \ \n'assetStatus': {'state': 'CREATING'}}\n"
      x-openapi-router-controller: swagger_server.controllers.sites_controller
  /sites/{siteId}:
    get:
      tags:
      - Sites
      operationId: get_site_by_id
      parameters:
      - name: siteId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        "200":
          description: "{'ResponseMetadata': {'RequestId': '87500d5d-9742-46f4-b3b5-8490405851dc',\n\
            \  'HTTPStatusCode': 200,\n  'HTTPHeaders': {'date': 'Mon, 28 Dec 2020\
            \ 14:43:56 GMT',\n   'content-type': 'application/json',\n   'content-length':\
            \ '1679',\n   'connection': 'keep-alive',\n   'x-amzn-requestid': '87500d5d-9742-46f4-b3b5-8490405851dc'},\n\
            \  'RetryAttempts': 0},\n  \n 'assetId': '0e6499c9-2945-4bd8-8c93-3e193602d0a3',\n\
            \ 'assetArn': 'arn:aws:iotsitewise:eu-west-1:241091466505:asset/0e6499c9-2945-4bd8-8c93-3e193602d0a3',\n\
            \ 'assetName': 'site333',\n 'assetModelId': 'a106a0e4-1d20-46dc-a7a3-8dd3f902f48d',\n\
            \ 'assetProperties': [\n   {'id': '08431a62-3bf0-4458-aac8-84f2452b5cca',\n\
            \   'name': 'SiteName',\n   'notification': {'topic': '$aws/sitewise/asset-models/a106a0e4-1d20-46dc-a7a3-8dd3f902f48d/assets/0e6499c9-2945-4bd8-8c93-3e193602d0a3/properties/08431a62-3bf0-4458-aac8-84f2452b5cca',\n\
            \    'state': 'DISABLED'},\n   'dataType': 'STRING'},\n  {'id': 'c3bac811-b33c-4c2c-b1f8-9c26c91d4274',\n\
            \   'name': 'Site Sensor Node Avg Temp (5min)',\n   'notification': {'topic':\
            \ '$aws/sitewise/asset-models/a106a0e4-1d20-46dc-a7a3-8dd3f902f48d/assets/0e6499c9-2945-4bd8-8c93-3e193602d0a3/properties/c3bac811-b33c-4c2c-b1f8-9c26c91d4274',\n\
            \    'state': 'DISABLED'},\n   'dataType': 'DOUBLE'},\n  {'id': '32c1459c-cb22-41ed-9b00-3de4699f4075',\n\
            \   'name': 'Site Asset Sensor Avg Temp (5min)',\n   'notification': {'topic':\
            \ '$aws/sitewise/asset-models/a106a0e4-1d20-46dc-a7a3-8dd3f902f48d/assets/0e6499c9-2945-4bd8-8c93-3e193602d0a3/properties/32c1459c-cb22-41ed-9b00-3de4699f4075',\n\
            \    'state': 'DISABLED'},\n   'dataType': 'DOUBLE'},\n  {'id': '5118bcc5-0117-4082-979c-98e24e4f8244',\n\
            \   'name': 'Site Avg Power Usage (5min)',\n   'notification': {'topic':\
            \ '$aws/sitewise/asset-models/a106a0e4-1d20-46dc-a7a3-8dd3f902f48d/assets/0e6499c9-2945-4bd8-8c93-3e193602d0a3/properties/5118bcc5-0117-4082-979c-98e24e4f8244',\n\
            \    'state': 'DISABLED'},\n   'dataType': 'DOUBLE'}],\n 'assetHierarchies':\
            \ [{'id': 'd883ac38-6437-4180-8ef5-124d0358241a',\n   'name': 'ZoneChilds'}],\n\
            \ 'assetCompositeModels': [],\n 'assetCreationDate': datetime.datetime(2020,\
            \ 12, 28, 12, 7, 26, tzinfo=tzlocal()),\n 'assetLastUpdateDate': datetime.datetime(2020,\
            \ 12, 28, 12, 7, 26, tzinfo=tzlocal()),\n 'assetStatus': {'state': 'ACTIVE'}}\n"
        "404":
          description: Entity with specified id was not found
      x-openapi-router-controller: swagger_server.controllers.sites_controller
    delete:
      tags:
      - Sites
      operationId: delete_site
      parameters:
      - name: siteId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        "204":
          description: "Entity removed successfully\n\n{'ResponseMetadata': {'RequestId':\
            \ '29725450-cf09-4a74-b510-a525c915834f', \n'HTTPStatusCode': 202, 'HTTPHeaders':\
            \ {'date': 'Mon, 28 Dec 2020 11:09:41 GMT', 'content-type': 'application/json',\
            \ \n'content-length': '36', 'connection': 'keep-alive', 'x-amzn-requestid':\
            \ '29725450-cf09-4a74-b510-a525c915834f'}, \n'RetryAttempts': 0}, \n'assetStatus':\
            \ {'state': 'DELETING'}}\n"
        "404":
          description: Entity with specified id was not found
        "406":
          description: Can not be delated because there are zones associated...
      x-openapi-router-controller: swagger_server.controllers.sites_controller
  /sites/properties/:
    put:
      tags:
      - Sites
      operationId: put_site_property_value
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetProperty'
        required: true
      responses:
        "200":
          description: |2

            {'ResponseMetadata': {'RequestId': 'e4209028-6785-4769-a639-ff51e6be8fb6',
            'HTTPStatusCode': 200,
            'HTTPHeaders': {'date': 'Mon, 28 Dec 2020 10:56:17 GMT',
             'content-type': 'application/json',
             'content-length': '19',
             'connection': 'keep-alive',
             'x-amzn-requestid': 'e4209028-6785-4769-a639-ff51e6be8fb6'},
            'RetryAttempts': 0},
            'errorEntries': []}
        "404":
          description: Entity with specified id was not found
      x-openapi-router-controller: swagger_server.controllers.sites_controller
  /sites/{siteId}/zones:
    get:
      tags:
      - Sites
      operationId: get_site_zones
      parameters:
      - name: siteId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        "200":
          description: "Get zones of the site\n\n{'ResponseMetadata': {'RequestId':\
            \ '618e1814-5af5-40d0-b9d0-c60f146cd7a1',\n  'HTTPStatusCode': 200,\n\
            \  'HTTPHeaders': {'date': 'Mon, 28 Dec 2020 14:25:22 GMT',\n   'content-type':\
            \ 'application/json',\n   'content-length': '483',\n   'connection': 'keep-alive',\n\
            \   'x-amzn-requestid': '618e1814-5af5-40d0-b9d0-c60f146cd7a1'},\n  'RetryAttempts':\
            \ 0},\n  \n 'assetSummaries': [\n    {'id': '5d363986-a624-4e2b-a122-3ef01d6cde78',\n\
            \   'arn': 'arn:aws:iotsitewise:eu-west-1:241091466505:asset/5d363986-a624-4e2b-a122-3ef01d6cde78',\n\
            \   'name': 'newzone333',\n   'assetModelId': '2bcfff28-ba76-44e5-8d8d-af10336fb448',\n\
            \   'creationDate': datetime.datetime(2020, 12, 28, 12, 41, 28, tzinfo=tzlocal()),\n\
            \   'lastUpdateDate': datetime.datetime(2020, 12, 28, 12, 41, 28, tzinfo=tzlocal()),\n\
            \   'status': {'state': 'ACTIVE'},\n   'hierarchies': [{'id': 'b02d7852-f3e1-432f-9942-b9906451f82c',\n\
            \     'name': 'SensorNodeChilds'},\n    {'id': '9f909579-8fef-4772-bbbb-a880018775ac',\n\
            \     'name': 'AssetSensorChilds'}]},\n     \n    {'id': '5d363986-a624-4e2b-a122-3ef01d6cde78',\n\
            \   'arn': 'arn:aws:iotsitewise:eu-west-1:241091466505:asset/5d363986-a624-4e2b-a122-3ef01d6cde78',\n\
            \   'name': 'newzone444',\n   'assetModelId': '2bcfff28-ba76-44e5-8d8d-af10336fb448',\n\
            \   'creationDate': datetime.datetime(2020, 12, 28, 12, 41, 28, tzinfo=tzlocal()),\n\
            \   'lastUpdateDate': datetime.datetime(2020, 12, 28, 12, 41, 28, tzinfo=tzlocal()),\n\
            \   'status': {'state': 'ACTIVE'},\n   'hierarchies': [{'id': 'b02d7852-f3e1-432f-9942-b9906451f82c',\n\
            \     'name': 'SensorNodeChilds'},\n    {'id': '9f909579-8fef-4772-bbbb-a880018775ac',\n\
            \     'name': 'AssetSensorChilds'}]}\n  ]\n}\n"
        "404":
          description: Entity with specified id was not found
      x-openapi-router-controller: swagger_server.controllers.sites_controller
  /sites/properties/{siteId}/:
    get:
      tags:
      - Sites
      operationId: get_site_property_values
      parameters:
      - name: siteId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        "200":
          description: "Get site property values\n\n{\n'SiteName': {'value': {'stringValue':\
            \ 'Site 0'}, 'timestamp': {'timeInSeconds': 1607011712, 'offsetInNanos':\
            \ 0}, 'quality': 'GOOD'}, \n'Site Sensor Node Avg Temp (5min)': {'value':\
            \ {'doubleValue': '23.0'}, 'timestamp': {'timeInSeconds': 1607011700,\
            \ 'offsetInNanos': 0}, 'quality': 'GOOD'}}, \n'Site Asset Sensor Avg Temp\
            \ (5min)': {'value': {'doubleValue': '23.0'}, 'timestamp': {'timeInSeconds':\
            \ 1607011700, 'offsetInNanos': 0}, 'quality': 'GOOD'}}, \n'Site Avg Power\
            \ Usage (5min)': {'value': {'doubleValue': '23.0'}, 'timestamp': {'timeInSeconds':\
            \ 1607011700, 'offsetInNanos': 0}, 'quality': 'GOOD'}}\n}\n\n# NOTE:\n\
            #  - if the property hasn't received any value yet from the sensor\n#\
            \    the response will be 'null' like this:\n  'Site Avg Power Usage (5min)':\
            \ 'null'\n"
        "404":
          description: Entity with specified id was not found
      x-openapi-router-controller: swagger_server.controllers.sites_controller
  /sites/properties/history:
    post:
      tags:
      - Sites
      operationId: get_site_property_value_history
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetPropertyValueHistoryRequest'
        required: true
      responses:
        "200":
          description: |
            Get property values from a specified time range.
        "404":
          description: Entity with specified id was not found
      x-openapi-router-controller: swagger_server.controllers.sites_controller
  /sites/{siteId}/associate/{zoneId}:
    post:
      tags:
      - Sites
      operationId: associate_zone
      parameters:
      - name: siteId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      - name: zoneId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        "200":
          description: |
            Associate a zone to a site

            {'ResponseMetadata': {'RequestId': 'f0d2f325-59fb-4b14-93e9-5a565383eae2',
            'HTTPStatusCode': 200,
            'HTTPHeaders': {'date': 'Mon, 28 Dec 2020 14:23:27 GMT',
             'content-type': 'application/json',
             'content-length': '0',
             'connection': 'keep-alive',
             'x-amzn-requestid': 'f0d2f325-59fb-4b14-93e9-5a565383eae2'},
            'RetryAttempts': 0}}
        "404":
          description: Entity with specified id was not found
      x-openapi-router-controller: swagger_server.controllers.sites_controller
  /sites/{siteId}/disassociate/{zoneId}:
    post:
      tags:
      - Sites
      operationId: disassociate_zone
      parameters:
      - name: siteId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      - name: zoneId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        "200":
          description: |
            Disssociate a zone from a site

            {'ResponseMetadata': {'RequestId': 'f0d2f325-59fb-4b14-93e9-5a565383eae2',
            'HTTPStatusCode': 200,
            'HTTPHeaders': {'date': 'Mon, 28 Dec 2020 14:23:27 GMT',
             'content-type': 'application/json',
             'content-length': '0',
             'connection': 'keep-alive',
             'x-amzn-requestid': 'f0d2f325-59fb-4b14-93e9-5a565383eae2'},
            'RetryAttempts': 0}}
        "404":
          description: Entity with specified id was not found
      x-openapi-router-controller: swagger_server.controllers.sites_controller
...
components:
  schemas:
...
    AssetProperty:
      type: object
      properties:
        propertyName:
          type: string
        assetId:
          type: string
        dataType:
          type: string
          enum: [ 'integerValue', 'stringValue', 'doubleValue', 'booleanValue' ]
        value:
          # TODO
          type: string
    Id:
      type: string
...
    NameObj:
      type: object
      properties:
        Name:
          type: string
...
    AssetPropertyValueHistoryRequest:
      type: object
      required:
      - assetId
      - startDate
      - endDate
      properties:
        assetId:
          type: string
        maxResults:
          type: number
          example: 50
        startDate:
          type: string
          format: datetime
          pattern: ([0-9]{4})-([0-9]{2})-([0-9]{2})
          example: "2020-12-20"
        endDate:
          type: string
          format: datetime
          pattern: ([0-9]{4})-([0-9]{2})-([0-9]{2})
          example: "2020-12-30"
        timeOrdering:
          type: string
          enum: ["ASCENDING", "DESCENDING"]
        include_prop:
          type: array
          items:
            type: string
          example: ["Temperature", "WorkingState"]


####################### SECURITY
  securitySchemes:
    basicAuth:     # <-- arbitrary name for the security scheme
      type: http
      scheme: basic
      x-basicInfoFunc: application.basic_auth
security:
  - basicAuth: ['secret']  # <-- use the same name here
#  - basicAuth: ['secret']  # <-- use the same name here

#   securitySchemes:
#     oAuth2Password:
#       type: oauth2
#       description: See https://developers.getbase.com/docs/rest/articles/oauth2/requests
#       flows:
#         password:
#           tokenUrl: https://api.getbase.com/oauth2/token
#           scopes: {}
# security:
#   - oAuth2Password: []


#  responses:
#    TODO!!!